Contains a refactored function for schedule seven day timer as well as hard corded constants in order to avoid magic strings
