+++ Section 8/HelperFiles/TimerController.swift
@@ -133,7 +133,7 @@ class TimerController {
     
 
     // This is a reUsable schedul Notifc alert however it needs to be worked on
-    func scheduleLocalNotificationOnTimer(identifier: String, actionTitle: String, categoryID: String, contentTitle: String, contentSubtitle: String, contentBody: String, contentBadge: NSNumber, contentSound: UNNotificationSound, contentLaunchImage: String, desiredTimeInterval: Int, resourceName: String, extenstionType: String) {
+func scheduleLocalNotificationOnTimer(identifier: String, actionTitle: String, categoryID: String, contentTitle: String, contentSubtitle: String, contentBody: String, contentBadge: NSNumber, contentSound: UNNotificationSound, contentLaunchImage: String, desiredTimeInterval: Int, resourceName: String, extenstionType: String) {
         
         // The Alert Button options
         let action = UNNotificationAction(identifier: identifier, title: actionTitle, options: [])
@@ -173,44 +173,7 @@ class TimerController {
     }
     
     
-    func scheduleLocalNotificationOnTimerUpdatedbyIvan(dissMissActionID: String, actionTitle: String, categoryID: String, contentTitle: String, contentSubtitle: String, contentBody: String, contentBadge: NSNumber, contentSound: UNNotificationSound, contentLaunchImage: String, desiredTimeInterval: Int, resourceName: String, extenstionType: String) {
-        
-        // The Alert Button options
-        let dissMissAction = UNNotificationAction(identifier: dissMissActionID, title: actionTitle, options: [])
-        let category = UNNotificationCategory(identifier: categoryID, actions: [dissMissAction], intentIdentifiers: [], options: [.customDismissAction])
-        UNUserNotificationCenter.current().setNotificationCategories([category])
-        
-        // MARK: - The notification payload
-        let content = UNMutableNotificationContent()
-        content.title = contentTitle
-        content.subtitle = contentSubtitle
-        content.body = contentBody
-        content.badge = contentBadge
-        content.sound = contentSound
-        content.launchImageName = contentLaunchImage
-        //This is the String ID for what we want to prseent to the user
-        content.categoryIdentifier = categoryID
-        /*When you schedule a notification request containing the attachment, the attachment’s file is moved to a new location to facilitate access by the appropriate processes. After the move, the only way to access the file is using the methods of the UNUserNotificationCenter object.*/
-        //        guard let url = Bundle.main.url(forResource: "homeFound", withExtension: "jpeg") else {return}
-        guard let url = Bundle.main.url(forResource: resourceName, withExtension: extenstionType) else {return}
-        do {
-            let attachments =  try UNNotificationAttachment(identifier: categoryID, url: url, options: [:])
-            
-            content.attachments = [attachments]
-        } catch {
-            print("\n\n🚀 There was an error with the attachment in: \(#file) \n\n \(#function); \n\n\(error); \n\n\(error.localizedDescription) 🚀\n\n")
-        }
-        
-        
-        let trigger = UNTimeIntervalNotificationTrigger(timeInterval: TimeInterval(desiredTimeInterval), repeats: true)
-        
-        let request = UNNotificationRequest(identifier: dissMissActionID, content: content, trigger: trigger)
-        UNUserNotificationCenter.current().add(request) { (error) in
-            if let error = error {
-                print("💩 There was an error in \(#function) ; (error) ; \(error.localizedDescription) 💩")
-            }
-        }
-    }
+    
     
     // MARK: - Calendar Notification
     func scheduleLocalNotificationOnDate(schedulateDateNotifId: String, dissmissActionTitle: String, editDateActionTitle: String, editDateOption: UNNotificationActionOptions,  categoryID: String, contentTitle: String, contentSubtitle: String, contentBody: String, contentBadge: NSNumber, contentSound: UNNotificationSound, contentLuanchImage: String, desiredTimeInterval: Int, resourceName: String, extenstionType: String, alrmComponent: Date) {
