+++ Section 8/Models/Section8Apartment.swift
@@ -8,17 +8,17 @@
 
 import Foundation
 
-class Section8Apartment: {
-    
-    var name: String
-    var address: String
-    var phone: String
-    var apartmentPhoto: String
-    
-    init(name: String, address: String, phone: String, apartmentPhoto: String) {
-        self.name = name
-        self.address = address
-        self.phone = phone
-        self.apartmentPhoto = apartmentPhoto
-    }
-}
+//class Section8Apartment: {
+//    
+//    var name: String
+//    var address: String
+//    var phone: String
+//    var apartmentPhoto: String
+//    
+//    init(name: String, address: String, phone: String, apartmentPhoto: String) {
+//        self.name = name
+//        self.address = address
+//        self.phone = phone
+//        self.apartmentPhoto = apartmentPhoto
+//    }
+//}
