+++ Section 8/ViewControllers/StepDetailDeleteOnceDone.swift
@@ -1,0 +1,22 @@
+//
+//  StepDetailDeleteOnceDone.swift
+//  Section 8
+//
+//  Created by Ivan Ramirez on 12/7/18.
+//  Copyright © 2018 Warren. All rights reserved.
+//
+
+import Foundation
+
+extension StepDetailTVC {
+//func scheduleSevenDayNotificationTestFunc() {
+//
+//    print("\n📅 7 day notification set\n")
+//    timerController.scheduleLocalNotificationOnTimer(identifier: sevenDayTimerID,actionTitle: NSLocalizedString("7DayDismiss", comment: ""), categoryID: categorySevenNotificationID, contentTitle: NSLocalizedString("7DayContentTitle", comment: ""), contentSubtitle: NSLocalizedString("7DayContentSubtitle", comment: ""), contentBody: NSLocalizedString("7DayContentBody", comment: ""), contentBadge: 1,contentSound: UNNotificationSound.default, contentLaunchImage: "", desiredTimeInterval: sevenDays, resourceName: NSLocalizedString("notificationBanner", comment: ""), extenstionType: "png")
+//}
+//
+//
+    
+    
+    
+}
